Implements delete, list, update and battle endpoints
The delete and update requires an existing transformer for the given id, if there is no transformer for the given id it returns an error (Not accepted).

There are 6 transformers created on startup, there inserted through h2 database which reads the "data.sql" file on the resources of the project.